feat: espresso verification with error messages (#82)
* feat: espresso verification with error messages

Motivation: it's hard to figure out why verifications fail. Currently we
compress all the verification results into a single boolean, thereby
discarding all information and making debugging unnecessarily hard for
us.

With this change our rust library returns an error message alongside the
boolean result (value set as before).

The idea is that wherever we call this code from go we can easily log
the error. The existing go code function wrappers should take care of
free-ing the memory allocated in rust but I don't know how to actually
test if this is working.

A bit of verification code has been re-organized to be able to create a
specific error message.

* add comment about allocating new string